agents/customer/pyproject.toml
[project]
name = "langchain-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "langchain-agent-server>=0.1.0",
    "asyncio>=4.0.0",
    "fastapi>=0.115.0",
    "fastmcp>=2.13.0.2",
    "httpx>=0.27.0",
    "langchain>=1.0.3",
    "langchain-core>=1.0.3",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.2",
    "langgraph-cli[inmem]>=0.4.7",
    "pyvault-agent>=0.2.0",
    "uvicorn>=0.32.0",
]
